Update
Added functionality for chaining equations, and making use of the +/- sign should the user wish to start with a negative integer. Moreover, some visual changes have been made (i.e replaced the C and CE buttons with one button which clears all entries made so far), but kept the back button should the user which to only remove one element.